+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Head
+
+- Fixed: added vendor namespace.
+
 # 1.0.0 - 2016-12-03
 
 - Chore: initial public release.